work on norm and exit codes, not finished
