+++ outfile.txt
@@ -1,3 +1,9 @@
-hi
-ho
-hu
+b
+v
+g
+f
+
+a
+b
+c
+d
